backend/requirements.txt
python-chess==1.999
tqdm==4.65.0
pygame==2.6.1
matplotlib==3.7.1
numpy==1.26.4
plotly==5.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.3.0
scikit-learn==1.2.2
PyQt5==5.15.11
Pillow==9.5.0
flask==2.3.3
flask-cors==4.0.0